Updated tutorial on data cleaning
Updated to 2.6.6

It was broght to my attention that the tutorial on data cleaning had typos.
It turned out that I hadn't updated this since the move from Python 2.7 to 3.

I went through that tutorial and updated code, and checked it still worked.
I removed some sections on the example of family L1872 on missing data in the candidates
and relatedness between top and second candidates, because this threw errors due to arrays
including/not including entries for missing fathers. This would be fixable, but there were
some horrid list comprehensions to parse, and I judged it not worth my time.

I also removed .ipynb files from /docs (the files that get compiled are in /docs/tutorials).
These were outdated files that I seem to have kept in case I still needed them, but it seems
that at least one user was attempting to learn from them.