2.7.4 Fixed bug in transition_probabilities
1st June 2023

Fixed a bug in transition_probability that underestimated probabilities of
missing heterozygous parents by a factor of two.

Set default prior on missing parents to 0.5 (i.e. missing and sampled parents
are equally likely, so the information comes through the data) because this
value is difficult to set unless you have good information ahead of time.

 On branch master
 Changes to be committed:
	modified:   faps/paternity_array.py
	modified:   faps/transition_probability.py
	modified:   setup.py